requirements.txt
torch>=2.1
transformers>=4.44.0
datasets>=2.20.0
accelerate>=0.33.0
peft>=0.12.0
trl>=0.9.0
pyyaml>=6.0
tqdm
trl>=0.9.0
